+++ src/main/java/chukhlantsev/oleg/taskboard/repository/UserRepository.java
@@ -2,15 +2,13 @@ package chukhlantsev.oleg.taskboard.repository;
 
 import chukhlantsev.oleg.taskboard.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
-
+/*
     @Query(value = """
             SELECT exists ( 
             SELECT 1 
@@ -21,6 +19,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
             """, nativeQuery = true)
     boolean isTaskOwner(@Param("userID") Long userID, @Param("taskID") Long taskID);
+*/
 
 
 }